# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Platform Email Transaksional \u0026 SMTP Delivery Profesional Indonesia.",
        "title": "RealSend API",
        "termsOfService": "https://realsend.web.id/legal/sla",
        "contact": {
            "name": "RealSend Support",
            "url": "https://realsend.web.id",
            "email": "support@realsend.web.id"
        },
        "version": "1.0"
    },
    "host": "localhost:3001",
    "basePath": "/api/v1",
    "paths": {
...
        "/emails/send": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mengirim email transaksional menggunakan domain yang terverifikasi dan aktif.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "summary": "Kirim email transaksional",
                "parameters": [
                    {
                        "description": "Payload data email",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_realsend_be-realsend_internal_service.SendEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Email masuk antrean (queued)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request atau domain tidak ditemukan/tidak terverifikasi",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "API Key tidak valid atau tidak disertakan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "Payload tidak valid (validasi gagal)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/emails/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mendapatkan detail status pengiriman email tertentu berdasarkan ID Log UUID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Emails"
                ],
                "summary": "Ambil status log email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email Log UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail log email",
                        "schema": {
                            "$ref": "#/definitions/github_com_realsend_be-realsend_internal_models.EmailLog"
                        }
                    },
                    "400": {
                        "description": "Format UUID tidak valid",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "API Key tidak valid atau tidak disertakan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Log email tidak ditemukan",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_realsend_be-realsend_internal_models.BounceType": {
            "type": "string",
            "enum": [
                "hard",
                "soft",
                "none"
            ],
            "x-enum-varnames": [
                "BounceHard",
                "BounceSoft",
                "BounceNone"
            ]
        },
        "github_com_realsend_be-realsend_internal_models.EmailLog": {
            "type": "object",
            "properties": {
                "api_key_id": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "bcc_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "bounce_reason": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "bounce_reason_str": {
                    "description": "Helper presentation strings",
                    "type": "string"
                },
                "bounce_type": {
                    "$ref": "#/definitions/github_com_realsend_be-realsend_internal_models.BounceType"
                },
                "cc_addresses": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "clicked_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "clicked_at_str": {
                    "type": "string"
                },
                "clicked_count": {
                    "type": "integer"
                },
                "content_type": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "delivered_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "delivered_at_str": {
                    "type": "string"
                },
                "domain_id": {
                    "$ref": "#/definitions/uuid.NullUUID"
                },
                "from_address": {
                    "type": "string"
                },
                "headers": {
                    "description": "Raw JSONB",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "id": {
                    "type": "string"
                },
                "metadata": {
                    "description": "Raw JSONB",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "opened_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "opened_at_str": {
                    "type": "string"
                },
                "opened_count": {
                    "type": "integer"
                },
                "queued_at": {
                    "type": "string"
                },
                "sent_at": {
                    "$ref": "#/definitions/sql.NullTime"
                },
                "sent_at_str": {
                    "type": "string"
                },
                "smtp_message_id": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "smtp_message_id_str": {
                    "type": "string"
                },
                "smtp_response": {
                    "$ref": "#/definitions/sql.NullString"
                },
                "smtp_response_str": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_realsend_be-realsend_internal_models.EmailStatus"
                },
                "subject": {
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "to_address": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_realsend_be-realsend_internal_models.EmailStatus": {
            "type": "string",
            "enum": [
                "queued",
                "processing",
                "sent",
                "delivered",
                "bounced",
                "rejected",
                "failed",
                "opened",
                "clicked"
            ],
            "x-enum-varnames": [
                "StatusQueued",
                "StatusProcessing",
                "StatusSent",
                "StatusDelivered",
                "StatusBounced",
                "StatusRejected",
                "StatusFailed",
                "StatusOpened",
                "StatusClicked"
            ]
        },
        "github_com_realsend_be-realsend_internal_service.SendEmailRequest": {
            "type": "object",
            "required": [
                "body",
                "from",
                "subject",
                "to"
            ],
            "properties": {
                "bcc": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "body": {
                    "type": "string"
                },
                "cc": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "content_type": {
                    "type": "string",
                    "enum": [
                        "text/plain",
                        "text/html"
                    ]
                },
                "from": {
                    "type": "string"
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "subject": {
                    "type": "string",
                    "maxLength": 998,
                    "minLength": 1
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "to": {
                    "type": "string"
                }
            }
        },
...
        "sql.NullString": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
        "sql.NullTime": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "uuid.NullUUID": {
            "type": "object",
            "properties": {
                "uuid": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if UUID is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key untuk integrasi aplikasi \u0026 kirim email",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Format: \"Bearer [JWT_TOKEN]\" untuk akses API dashboard",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}